requirements.txt
groq>=0.8.0
PyPDF2==3.0.1
python-dotenv==1.0.0
